WebCL-conformance/conformance/extension/KHR_gl_sharing/cl_memoryObject_getGLObjectInfo.html

<!--
 Copyright (C) 2014 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLMemoryObject getGLObjectInfo conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
    <script src="../../../resources/clgl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<div style="position:absolute; left:0px; top:0px; visibility:hidden" id="x">
    <canvas id="inputCanvas" />
</div>
<script>
try {
    debug("WebCLMemoryObject getGLObjectInfo test case.");

    var wtu = WebCLTestUtils;
    var clgl = WebCLGLTestUtils;
    var canvas = document.getElementById("inputCanvas");
    var glContext = canvas.getContext("experimental-webgl");

    if (!wtu.enableExtension(webcl, "KHR_gl_sharing")) {
        throw { description : "Failed to enable extension : KHR_gl_sharing. Exitting..."};
    }

    var webCLPlatform = wtu.getPlatform();
    var webCLDevices = wtu.getDevices(webCLPlatform);
    var webCLGLContext = wtu.createContext(glContext, webCLDevices);

    var glBuffer = clgl.createBuffer(glContext);
    clgl.bindBuffer(glContext, glContext.ARRAY_BUFFER, glBuffer);
    var webCLBufferMemoryObject = clgl.createFromGLBuffer(webCLGLContext, webcl.MEM_READ_ONLY, glBuffer);
    shouldBeType("webCLBufferMemoryObject.getGLObjectInfo();", "WebCLGLObjectInfo");
    shouldBeType("webCLBufferMemoryObject.getGLObjectInfo().glObject;", "WebGLBuffer");
    shouldEvaluateTo("webCLBufferMemoryObject.getGLObjectInfo().type", "webcl.GL_OBJECT_BUFFER");
    shouldBeUndefined("webCLBufferMemoryObject.getGLObjectInfo().textureTarget;");
    shouldBeUndefined("webCLBufferMemoryObject.getGLObjectInfo().mipmapLevel;");

    var glRenderBuffer = clgl.createRenderbuffer(glContext);
    clgl.bindRenderbuffer(glContext, glContext.RENDERBUFFER, glRenderBuffer);
    var webCLRenderbufferMemoryObject = clgl.createFromGLRenderbuffer(webCLGLContext, webcl.MEM_READ_ONLY, glRenderBuffer);
    shouldBeType("webCLRenderbufferMemoryObject.getGLObjectInfo();", "WebCLGLObjectInfo");
    shouldBeType("webCLRenderbufferMemoryObject.getGLObjectInfo().glObject;", "WebGLRenderbuffer");
    shouldEvaluateTo("webCLRenderbufferMemoryObject.getGLObjectInfo().type;", "webcl.GL_OBJECT_RENDERBUFFER");
    shouldBeUndefined("webCLRenderbufferMemoryObject.getGLObjectInfo().textureTarget;");
    shouldBeUndefined("webCLRenderbufferMemoryObject.getGLObjectInfo().mipmapLevel;");

    var glTexture = clgl.createTexture(glContext);
    clgl.bindTexture(glContext, glContext.TEXTURE_2D, glTexture);
    var miplevel = 1;
    var webCLTextureMemoryObject = clgl.createFromGLTexture(webCLGLContext, webcl.MEM_READ_ONLY, glContext.TEXTURE_2D, miplevel, glTexture);
    shouldBeType("webCLTextureMemoryObject.getGLObjectInfo();", "WebCLGLObjectInfo");
    shouldBeType("webCLTextureMemoryObject.getGLObjectInfo().glObject;", "WebGLTexture");
    shouldEvaluateTo("webCLTextureMemoryObject.getGLObjectInfo().type;", "webcl.GL_OBJECT_TEXTURE2D");
    shouldEvaluateTo("webCLTextureMemoryObject.getGLObjectInfo().textureTarget;", "glContext.TEXTURE_2D");
    shouldEvaluateTo("webCLTextureMemoryObject.getGLObjectInfo().mipmapLevel;", "miplevel");

} catch(e) {
    testFailed(e.description);
}
successfullyParsed = true;
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
